% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIDC.R
\name{.puc_per_target}
\alias{.puc_per_target}
\title{I_XiZ, Ispec_XiZ and p_Z is list with equally length, theoretically, p_Z is the same value.}
\usage{
.puc_per_target(I_XiZ, Ispec_XiZ, p_Z)
}
\description{
I_XiZ, Ispec_XiZ and p_Z is list with equally length, theoretically, p_Z is the same value.
}
