% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIDC.R
\name{matToNet}
\alias{matToNet}
\title{Inferring regulatory network with PIDC matrix}
\usage{
matToNet(
  weightMat,
  methods = c("aracne", "clr", "mrnet", "mrnetb"),
  cutoff = NULL,
  ncores = 1
)
}
\arguments{
\item{weightMat}{A square matrix whose element should be positive values.}

\item{methods}{The name of the network inference algorithm. Default: aracne.}

\item{cutoff}{Set the cutoff of regulatory networks. Default: NULL, means
inferring cutoff by itself.}

\item{ncores}{Number of cores to use. Default:1.}
}
\value{
A data.frame or a list with regulatory networks.
}
\description{
Inferring gene regulatory network using weighted square matrix from PIDC output.
}
\examples{
data(expMat)
PIDC_res <- PIDC(expMat)
PIDC_net <- matToNet(PIDC_res)
}
