% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIDC.R
\name{run_PIDC}
\alias{run_PIDC}
\title{Inferring Gene Regulatory Network (GRN) using partial information decomposition and context (PIDC) method}
\usage{
run_PIDC(
  expMat,
  regulators = NULL,
  targets = NULL,
  logNormalized = FALSE,
  ncores = 1,
  verbose = interactive()
)
}
\arguments{
\item{expMat}{A \code{matrix} storing the gene expression data. Rows corresponding
to features (eg. gene symbols) and columns corresponding to cells. Raw read counts,
UMIs, TPMs or logNormalized counts were supported.}

\item{regulators}{The regulator genes used for GRN inferring (eg, transcription
factors). At least two regulators required. Default: NULL, using all the genes
in gene expression matrix.}

\item{targets}{The target genes used for GRN inferring. Default: NULL, using
all the genes in gene expression matrix.}

\item{logNormalized}{Whether the input have been logNormalized?}

\item{ncores}{Number of cores used for parallel calculation. The running time
heavily depend on the number of regulators and targets, for genes over 5000,
we strongly suggest you to use multi-cores. Default: 1.}

\item{verbose}{Whether to print message while running. Default: TRUE.}
}
\value{
A matrix with weighted value between regulatro and targets.
}
\description{
Inferring Gene Regulatory Network (GRN) using partial information decomposition and context (PIDC) method
}
\examples{
data(expMat)
PIDC_res <- PIDC(expMat)
head(PIDC_res)



}
