% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIDC.R
\name{matToNet}
\alias{matToNet}
\title{Get regulatory network with PIDC matrix}
\usage{
matToNet(
  mat,
  methods = c("aracne", "clr", "mrnet", "mrnetb"),
  return_regulons = FALSE,
  withWeight = FALSE
)
}
\arguments{
\item{mat}{A square matrix whose element is positive values.}

\item{methods}{The name of the network inference algorithm. Default: aracne.}

\item{return_regulons}{Whether to return a regulon list? Default: FALSE}

\item{withWeight}{The regulons is wight?}
}
\value{
A data.frame or a list with regulatory networks.
}
\description{
Inferring gene regulatory network using weighted square matrix from PIDC output.
}
\examples{
data(expMat)
PIDC_res <- PIDC(expMat)
net <- matToNet(mat=PIDC_res)

}
